Split metrics into separate files
No functional changes
